refactor: modularize WASM loading
- extract shared instantiation logic into wasmCore
- add browser and Node-specific WASM loaders
- use package import conditions to resolve the correct loader
- remove environment detection from evaluate.js
